Move the authentication part of the process inside companion app module
