+++ authentication/src/main/java/net/bonysoft/doityourselfie/authentication/AuthenticationListener.kt
@@ -1,6 +1,5 @@
 package net.bonysoft.doityourselfie.authentication
 
-import com.google.api.client.googleapis.extensions.android.gms.auth.UserRecoverableAuthIOException
 
 interface AuthenticationListener {
 
@@ -8,5 +7,5 @@ interface AuthenticationListener {
 
     fun showLoggedOutUi()
 
-    fun onUserRecoverableException(e: UserRecoverableAuthIOException)
+    fun onUserRecoverableException(e: Exception)
 }
