+++ companion/src/main/AndroidManifest.xml
@@ -1,2 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="net.bonysoft.doityourselfie.companion" />
+    package="net.bonysoft.doityourselfie.companion">
+
+    <application>
+        <activity
+            android:name=".MainActivity"
+            android:label="@string/title_activity_main"
+            android:theme="@style/AppTheme.NoActionBar"></activity>
+    </application>
+
+</manifest>
