+++ companion/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">DIY - Companion</string>
+    <string name="title_activity_main">MainActivity</string>
 </resources>
