+++ mobile/build.gradle
@@ -56,11 +56,6 @@ dependencies {
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation project(":secrets")
     implementation project(":photos")
-    implementation project(":authentication")
-
-    implementation 'com.google.android.gms:play-services-auth:15.0.1'
-    implementation 'com.google.api-client:google-api-client:1.22.0'
-    implementation 'com.google.api-client:google-api-client-android:1.22.0'
 
     implementation 'com.jakewharton.timber:timber:4.7.1'
     implementation 'com.squareup.picasso:picasso:2.71828'
